ahora las subidas de archivos tienen estan ligados a un usuario, agregue seguridad al enviar y borrar archivos

+++ controllers/storage.js
@@ -3,7 +3,7 @@ const { matchedData } = require('express-validator');
 const { storageModel } = require('../models');
 const { handleHttpError } = require('../utils/handleError');
 
-const PUBLIC_URL = (process.env.NODE_ENV === 'production') ? process.env.PUBLIC_URL_PROD : process.env.PUBLIC_URL_DEV 
+const PUBLIC_URL = (process.env.NODE_ENV === 'production') ? process.env.PUBLIC_URL_PROD : process.env.PUBLIC_URL_DEV
 const MEDIA_PATH = `${__dirname}/../storage`;
 /**
  * Obtener la base de datos!
@@ -54,8 +54,10 @@ const createItem = async (req, res) => {
         const fileData = {
             id: file.filename.split('.').shift(),
             filename: file.filename,
-            url:`${PUBLIC_URL}/${file.filename}`,
-            originalname: file.originalname.split('.').shift()
+            url: `${PUBLIC_URL}/${file.filename}`,
+            originalname: file.originalname.split('.').shift(),
+            ext: file.filename.split('.').pop(),
+            user_id: req.session.user.id
 
         };
 
@@ -63,7 +65,7 @@ const createItem = async (req, res) => {
 
         res.status(200).send({ data });
 
-    }catch(error) {
+    } catch (error) {
         console.error(error);
         handleHttpError(res, 'Error al guardar registro de archivo');
     }
@@ -76,20 +78,25 @@ const createItem = async (req, res) => {
 */
 const deleteItem = async (req, res) => {
     try {
-        const { id } = matchedData(req)
+        const { id } = matchedData(req);
         const dataFile = await storageModel.findByPk(id);
         if (!dataFile) {
-            return  handleHttpError(res, 'id no encontrado');
-        } else{
-            const { filename } = dataFile;
-            const filePath = `${MEDIA_PATH}/${filename}`;
-            const data = { filePath, deleted: 1 };
-            fs.unlinkSync(filePath);
-            await dataFile.destroy();
-            res.status(200).send({ data });
+            return handleHttpError(res, 'id no encontrado');
+        } else {
+            if (dataFile.user_id === req.session.user.id || (req.session.roles).includes(5)) {
+                const { filename } = dataFile;
+                const filePath = `${MEDIA_PATH}/${filename}`;
+                const data = { filePath, deleted: 1 };
+                fs.unlinkSync(filePath);
+                await dataFile.destroy();
+                res.status(200).send({ data });
+            }else{
+                return handleHttpError(res, 'DONT_HAVE_PERMISSION_TO_DELETE',401);
+            }
+
         }
-        
-        
+
+
     } catch (error) {
         console.error(error);
         handleHttpError(res, 'Error al intentar eliminar el archivo');
